backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.2